+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact project, newest changes on the top.
 
+## [Unrelease]
+
+### Fixed
+
+- `ui/Scroller` to support `data-webos-voice-focused`, `data-webos-voice-disabled`, `data-webos-voice-group-label` props
+
 ## [3.3.1] - 2020-07-20
 
 ### Changed